conclusao da reserva.

+++ Comandas.Api/Comandas.Api/Controllers/ReservasController.cs
@@ -1,7 +1,6 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Comandas.Api.Models;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
-using Comandas.Api;
-using Comandas.Api.Models;
 
 namespace Comandas.Api.Controllers
 {
@@ -38,7 +37,6 @@ namespace Comandas.Api.Controllers
         }
 
         // PUT: api/Reservas/5
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
         public async Task<IActionResult> PutReserva(int id, Reserva reserva)
         {
@@ -46,8 +44,25 @@ namespace Comandas.Api.Controllers
             {
                 return BadRequest();
             }
-
             _context.Entry(reserva).State = EntityState.Modified;
+
+            // -----------
+            var novaMesa = await _context.Mesas
+                    .FirstOrDefaultAsync(m => m.NumeroMesa == reserva.NumeroMesa);
+            if (novaMesa is null)
+                return BadRequest("Mesa não encontrada.");
+            novaMesa.SituacaoMesa = (int)SituacaoMesa.Reservada; // novaMesa agora está reservada
+
+            // consulta dados da reserva original
+            var reservaOriginal = await _context.Reservas.AsNoTracking()
+                .FirstOrDefaultAsync(r => r.Id == id);
+            // consulta numero da mesa original
+            var numeroMesaOriginal = reservaOriginal!.NumeroMesa;
+            // consulta a mesa original
+            var mesaOriginal = await _context.Mesas
+                    .FirstOrDefaultAsync(m => m.NumeroMesa == numeroMesaOriginal);
+            mesaOriginal!.SituacaoMesa = (int)SituacaoMesa.Livre; // mesa original agora está livre
+            // -----------
 
             try
             {
@@ -74,6 +89,27 @@ namespace Comandas.Api.Controllers
         public async Task<ActionResult<Reserva>> PostReserva(Reserva reserva)
         {
             _context.Reservas.Add(reserva);
+
+            // ------------
+            // consultando a mesa pelo numero
+            var mesa = await _context.Mesas
+                    .FirstOrDefaultAsync(m => m.NumeroMesa == reserva.NumeroMesa);
+
+            if (mesa is null)
+                return BadRequest("Mesa não encontrada.");
+
+            // se mesa encontrada
+            if (mesa is not null)
+            {
+                if (mesa.SituacaoMesa != (int)SituacaoMesa.Livre)
+                {
+                    return BadRequest("Mesa não está disponível para reserva.");
+                }
+
+                // atualizar o status da mesa para reservado
+                mesa.SituacaoMesa = (int)SituacaoMesa.Reservada;
+            }
+            // ------------
             await _context.SaveChangesAsync();
 
             return CreatedAtAction("GetReserva", new { id = reserva.Id }, reserva);
@@ -86,14 +122,23 @@ namespace Comandas.Api.Controllers
             var reserva = await _context.Reservas.FindAsync(id);
             if (reserva == null)
             {
-                return NotFound();
+                return NotFound("Reserva nao encontrada");
             }
+            // ------------
+            var mesa = await _context.Mesas
+                    .FirstOrDefaultAsync(m => m.NumeroMesa == reserva.NumeroMesa);
+            if (mesa is null)
+                return BadRequest("Mesa não encontrada.");
 
+            mesa.SituacaoMesa = (int)SituacaoMesa.Livre; // (int) converte o enum para int
+            // ------------
             _context.Reservas.Remove(reserva);
             await _context.SaveChangesAsync();
 
             return NoContent();
         }
+
+
 
         private bool ReservaExists(int id)
         {
